refactor: split state log and ui types
