+++ AGENTS.md
@@ -33,6 +33,8 @@ cargo test --features test-utils                         # Run all tests includi
 | `src/managers/ios/*.rs`       | iOS helper modules (`details/discovery/lifecycle` and upcoming splits)                   |
 | `src/managers/common.rs`      | `DeviceManager` trait                                                                    |
 | `src/app/api_levels.rs`       | API level management mode handling and install/uninstall orchestration                   |
+| `src/app/state/ui.rs`         | UI-facing state enums and confirmation dialog structs                                    |
+| `src/app/state/logs.rs`       | Log entry state and log-related `AppState` methods                                       |
 | `src/ui/render.rs`            | Three-panel layout rendering and panel composition                                       |
 | `src/ui/dialogs.rs`           | Modal dialog rendering (`create/confirm/api-level/notifications`)                        |
 | `src/ui/panels.rs`            | Panel rendering helpers (`android/ios/details/logs/commands`)                            |
